Move get_client_ip to membership.utils
